+++ CHANGELOG.md
@@ -17,11 +17,17 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## [7.28.0] - 2024-03-11
+## [7.28.0] - 2024-03-13
 ### Added
-- Added support for advanced filter query in the `list()` (and `__call__()`) method of `assets`, `events`, `sequences`, and `time_series` APIs. Now you are able to use advanced filter (like in `filter()`) at the same time as the simple filter properties allowing for more complex requests. Additionally, added `sort` and `partitions` params to `list()` and `__call__()` methods of all the APIs that were missing either.
+- Added support for advanced filter query in the `list()` (and `__call__()`) method of `assets`, `events`, `sequences`,
+  and `time_series` APIs. Now you are able to use advanced filter (like in `filter()`) at the same time as the simple
+  filter properties, allowing for more complex requests.
+- Added missing `sort` parameter to `list()` and `__call__()` methods for `AssetsAPI`.
+- Added missing `sort` parameter to `list()` and `__call__()` methods for `TimeSeriesAPI`.
+- Added missing `sort` and `partitions` parameters to `list()` and `__call__()` methods for `SequencesAPI`.
 ### Deprecated
-- Added a deprecation warning on the `filter()` method of `assets`, `events`, `sequences`, and `time_series` APIs as its functionality is fully replicated in the `list()` method.
+- Added a deprecation warning on the `filter()` method of `assets`, `events`, `sequences`, and `time_series` APIs as
+  its functionality is fully covered by the `list()` method.
 
 ## [7.27.2] - 2024-03-08
 ### Added